Update Binary Classification.py
Added pickle dump of logistic regression model at end of code.

Added Train validation split that uses random_state = None to create a few confusion matrices. Allows us to tune the hyperparameters of tfidf and logistic regression model and once we set the parameters we want, we train the model on all of the training data.